src/app/services/logInterceptor/log-interceptor.service.ts
import { Injectable } from '@angular/core';
import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent, HttpErrorResponse } from '@angular/common/http';
import { Observable, throwError } from 'rxjs';
import { catchError } from 'rxjs/operators';

@Injectable({
  providedIn: 'root'
})
export class LogInterceptorService implements HttpInterceptor {

  constructor() { }

  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
    const cloned = request.clone({params: request.params.append('client_id', 'y9xkxSl71A-5hTo2N_Dn8gbNqCU-UjRY0Y_yNyq3QHQ')});
    return next.handle(cloned);
  }

}
